pyproject.toml
[tool.poetry]
name = "guss"
version = "0.1.0"
description = ""
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = {extras = ["email"], version = "^1.9.1"}
jwcrypto = "^1.3.1"
PyYAML = "^6.0"
Django = "^4.0.5"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
